Add operator export command to expose metrics.
This will allow running operator pod as side-car container.
